till now

+++ sheet_1/d-difference.cpp
@@ -2,10 +2,10 @@
 using namespace std;
 
 int main() {
-    int a, b,c,d;
+    long long a, b,c,d;
     cin >> a >> b >> c >> d;
 
-    int ans = ((a*b)-(c*d));
+    long long ans = ((a*b)-(c*d));
     cout << "Difference = " << ans << endl;
     return 0;
 }
